+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 - `deprecated-color-setting-shorthand` auto-fixable diagnostic for replacing [deprecated shorthands for setting colors](https://github.com/marp-team/marpit/issues/331) ([#358](https://github.com/marp-team/marp-vscode/issues/358), [#366](https://github.com/marp-team/marp-vscode/pull/366))
+- Experimental `markdown.marp.strictPathResolutionDuringExport` setting ([#367](https://github.com/marp-team/marp-vscode/pull/367))
 - Sponsor button for Visual Studio Marketplace ([#365](https://github.com/marp-team/marp-vscode/pull/365))
 
 ## v2.0.1 - 2022-06-06